UserNavbarをt関数を使い、menuを分離
