+++ components/organisms/navbar/UserNavbar.vue
@@ -15,74 +15,25 @@
 
     <v-btn
       :to="{ name: 'users.storages.index' }"
+      v-text="$t('look_at_the_storage')"
       dark
       large
       class="d-none d-md-flex"
-    >
-      作品を眺める
-    </v-btn>
+    />
 
     <v-btn
       :to="{ name: 'users.storages.create' }"
+      v-text="$t('add_new_work')"
       dark
       large
       class="d-none d-md-flex"
-    >
-      新規作品
-    </v-btn>
+    />
 
     <locale-dropdown dark color="grey darken-4" />
 
     <!-- auth -->
     <template v-if="user">
-      <v-menu bottom offset-y>
-        <template v-slot:activator="{ on }">
-          <v-btn v-on="on" icon>
-            <v-icon>mdi-dots-vertical</v-icon>
-          </v-btn>
-        </template>
-
-        <v-list>
-          <v-list-item :to="{ name: 'users.storages.index' }">
-            <v-list-item-icon>
-              <v-icon>mdi-book-open-variant</v-icon>
-            </v-list-item-icon>
-            <v-list-item-content>
-              <v-list-item-title v-text="`作品を眺める`" />
-            </v-list-item-content>
-          </v-list-item>
-
-          <v-list-item :to="{ name: 'users.storages.create' }">
-            <v-list-item-icon>
-              <v-icon>mdi-pencil</v-icon>
-            </v-list-item-icon>
-            <v-list-item-content>
-              <v-list-item-title v-text="`新規作品`" />
-            </v-list-item-content>
-          </v-list-item>
-
-          <!-- ログアウト -->
-          <v-list-item @click.prevent="logout">
-            <v-list-item-icon>
-              <v-icon>mdi-account-off</v-icon>
-            </v-list-item-icon>
-            <v-list-item-content>
-              <v-list-item-title v-text="$t('logout')" />
-            </v-list-item-content>
-          </v-list-item>
-        </v-list>
-      </v-menu>
-    </template>
-
-    <!-- Guest -->
-    <template v-else>
-      <v-btn :to="{ name: 'login' }" text>
-        {{ $t('login') }}
-      </v-btn>
-
-      <v-btn :to="{ name: 'register' }" text>
-        {{ $t('register') }}
-      </v-btn>
+      <user-navbar-menu />
     </template>
   </v-app-bar>
 </template>
@@ -90,10 +41,12 @@
 <script>
 import { mapGetters } from 'vuex'
 import LocaleDropdown from '@/components/organisms/dropdown/LocaleDropdown'
+import UserNavbarMenu from '@/components/organisms/menu/UserNavbarMenu'
 
 export default {
   components: {
-    LocaleDropdown
+    LocaleDropdown,
+    UserNavbarMenu
   },
 
   data: () => ({
@@ -106,14 +59,6 @@ export default {
   }),
 
   methods: {
-    async logout() {
-      // Log out the user.
-      await this.$store.dispatch('auth/logout')
-
-      // Redirect to login.
-      this.$router.push({ name: 'login' })
-    },
-
     drawerEvent() {
       this.drawer = !this.drawer
       this.$emit('drawer', this.drawer)
